+++ ergo-lib/src/ast/expr.rs
@@ -3,13 +3,13 @@ use core::fmt;
 use crate::serialization::op_code::OpCode;
 use crate::types::stype::SType;
 
+use super::bin_op::BinOp;
 use super::coll_fold::Fold;
 use super::constant::Constant;
 use super::constant::ConstantPlaceholder;
 use super::extract_reg_as::ExtractRegisterAs;
 use super::global_vars::GlobalVars;
 use super::method_call::MethodCall;
-use super::ops;
 use super::option_get::OptionGet;
 use super::predef_func::PredefFunc;
 use super::property_call::PropertyCall;
@@ -37,7 +37,7 @@ pub enum Expr {
     /// Property call
     ProperyCall(Box<PropertyCall>),
     /// Binary operation
-    BinOp(ops::BinOp, Box<Expr>, Box<Expr>),
+    BinOp(Box<BinOp>),
     /// Option get method
     OptionGet(Box<OptionGet>),
     /// Extract register's value (box.RX properties)
@@ -56,6 +56,7 @@ impl Expr {
             Expr::Context => OpCode::CONTEXT,
             Expr::OptionGet(v) => v.op_code(),
             Expr::ExtractRegisterAs(v) => v.op_code(),
+            Expr::BinOp(op) => op.op_code(),
             _ => todo!("{0:?}", self),
         }
     }